test(repository): move integration `resolve{Relation}Metadata` tests to unit tests
