+++ packages/repository/src/__tests__/unit/repositories/relations-helpers/resolve-has-many-metadata.unit.ts
@@ -9,8 +9,8 @@ import {
   HasManyDefinition,
   ModelDefinition,
   RelationType,
-} from '../../..';
-import {resolveHasManyMetadata} from '../../../relations/has-many/has-many.helpers';
+} from '../../../..';
+import {resolveHasManyMetadata} from '../../../../relations/has-many/has-many.helpers';
 
 describe('resolveHasManyMetadata', () => {
   it('throws if the wrong metadata type is used', async () => {
